requirements.txt
Flask==3.0.0                                                                                                           
flask-cors==4.0.0                                                                                                      
python-dotenv==1.0.0                                                                                                   
cryptography==41.0.7                                                                                                   
supabase==2.3.0                                                                                                        
postgrest==0.13.0                                                                                                      
PyJWT==2.8.0                                                                                                           
qrcode[pil]==7.4.2                                                                                                     
python-dateutil==2.8.2                                                                                                 
pytest==7.4.0   